Added states FilteringExistingFiles and MakingIncrementalUpdates
Added extended state field PackageExists to see if this is a
regeneration. If package exists we need to filter files and do the
incremental updates. If packages does not exist we write them all.

We now return an error if either go.mod or main.go exists when
generation is run with the --init flag set.

Implemented FilteringExistingFiles